chore(accounts-endpoint): re-organize test folder structure
